Add files via upload
Concepts. Spatial Dominion is the one you heard tonight, Success/Fail Condition go with it. I was thinking of building on Faster and Faster for the actual song, probably chip but I can potentially build it up more with instruments. And make an elevator remix. Even if the main is chip. Just because.